backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
pypdf
python-multipart
sse-starlette
google-genai
langchain
langchain-google-genai
python-dotenv
